edit + dashboard

+++ src/main/java/com/dev/delta/repositories/CustomerRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.dev.delta.entities.Customer;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
-	Optional<Customer> findById(int id);
+	Optional<Customer> findById(Long id);
 }
