+++ src/main/java/com/dev/delta/repositories/RoomRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.dev.delta.entities.Room;
 
 public interface RoomRepository extends JpaRepository<Room, Long>  {
-	Optional<Room> findById(int id);
+	Optional<Room> findById(Long id);
 }
